FE: Move loading form
